+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - [ ] Indication of rules not matching any files
 - [ ] Validate syntax of sections
 - [ ] Swith icons to the official .editorconfig icon
-- [x] Indication of a rules that are already defined by parent .editorconfig
+- [x] Indication of rules that are already defined by parent .editorconfig
 - [x] Update tooltip to display larger icon
 - [x] Removing empty lines when sorting sections